Begin writing actual client generation code
